Have aibs-informatics-cdk-lib pkg indicate that it is typed
Currently, due to a missing `py.typed` file, this package will
will raise mypy typing errors when used as a dependency in
some other package.

This commit fixes this issue by adding the missing file and
updating the `pyproject.toml` to include the `py.typed` file
as part of `package_data`.

For details see:
https://mypy.readthedocs.io/en/stable/installed_packages.html#creating-pep-561-compatible-packages